Fix albumratings does not exist
The first time rating an album, a FileNotFoundError is raised
because ~/.config/clerk/albumratings.json does not exists:
- Moved the file opening (i.e. load_fastlist) inside the 'try'
block and added an except clause, to catch the FileNotFoundError
exception and process it accordingly.
- Moved the action of saving the json track database and setting the
sticker into a 'finally' block in order not to duplicate lines of code
in the different blocks of the try/except.
- Replaced the 'artist' tag with the 'albumartist' one.